+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * UI tests replacement with RTL/Jest for component `MultipieceModal`. Refs UICHKIN-283.
 * Cover `PrintButton` component by RTL/jest tests. Refs UICHKIN-285.
 * Cover `RouteForDeliveryModal` component by RTL/jest tests. Refs UICHKIN-286.
+* Add Jest/RTL testing for `ModalManager` component. Refs UICHKIN-288.
 
 ## [7.0.1] (https://github.com/folio-org/ui-checkin/tree/v7.0.1) (2022-04-06)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v7.0.0...v7.0.1)